feat(标准产监/模拟数据)：v1.1需求版本评价分析、经济效益、大数据模块模拟数据改造

+++ base/base-common/src/main/java/com/yjtech/wisdom/tourism/common/constant/DecisionSupportConstants.java
@@ -59,12 +59,12 @@ public class DecisionSupportConstants {
     /**
      * 模拟数据
      */
-    public static final Integer MOCK = 1;
+    public static final Byte MOCK = 1;
 
     /**
      * 真实数据
      */
-    public static final Integer IMPL = 0;
+    public static final Byte IMPL = 0;
 
     /**
      * 应急事件-同比
