+++ business/business-country-tour/src/main/java/com/yjtech/wisdom/tourism/vo/MonthPassengerFlowVo.java
@@ -43,5 +43,5 @@ public class MonthPassengerFlowVo extends TimeBaseQuery implements Serializable 
     /**
      * is_simulation 是否有模拟数据 1有 0无
      */
-    private Integer isSimulation = 0;
+    private Byte isSimulation = 0;
 }
