+++ business/business-decision-support/src/main/java/com/yjtech/wisdom/tourism/decisionsupport/business/strategyimpl/ProvinceOutsideTourStrategyImpl.java
@@ -32,7 +32,7 @@ public class ProvinceOutsideTourStrategyImpl extends BaseStrategy {
      * @return
      */
     @Override
-    public DecisionWarnEntity init(DecisionEntity entity, Integer isSimulation) {
+    public DecisionWarnEntity init(DecisionEntity entity, Byte isSimulation) {
 
         DecisionWarnEntity result = JSONObject.parseObject(JSONObject.toJSONString(entity), DecisionWarnEntity.class);
